Improved 'gae' admin command (added new actions).

+++ kalapy/admin/commands/project.py
@@ -33,6 +33,9 @@ def copy_template(template, target, context):
         if srcname.endswith('_t'):
             dstname = dstname[:-2]
 
+        if os.path.exists(dstname):
+            continue
+
         shutil.copy2(srcname, dstname)
 
         if srcname.endswith('_t'):
@@ -97,12 +100,47 @@ class StartApp(Command):
             os.mkdir('%s/%s' % (name, d))
 
 class GAEProject(ActionCommand):
-    """make this project google appengine compatible.
+    """Perform google appengine specific tasks.
+
+    Use the 'appcfg.py' and 'dev_appserver.py' (included in appengine sdk)
+    for other appengine specific tasks.
     """
     name = "gae"
 
-    def action_app(self, options, args):
-        """create appengine specific files (app.yaml, gaehandler.py).
+    options = (
+        ('a', 'address', 'localhost', 'hostname for the appserver'),
+        ('p', 'port', '8080', 'port number for the appserver'),
+        ('i', 'install', False, 'install libs (extra libs as arguments)'),
+    )
+
+    def setup_stubs(self):
+        from kalapy.db.engines.gae._utils import setup_stubs
+        setup_stubs()
+
+    def action_runserver(self, options, args):
+        """launch 'dev_appserver.py runserver'
+        """
+        self.setup_stubs()
+        from google.appengine.tools import dev_appserver_main
+        dev_appserver_main.main(
+            ['runserver', '-a', options.address, '-p', options.port, '.'])
+
+    def action_update(self, options, args):
+        """launch 'appcfg.py update'
+        """
+        self.setup_stubs()
+        from google.appengine.tools import appcfg
+        appcfg.main(['', 'update', '.'])
+
+    def action_rollback(self, options, args):
+        """launch 'appcfg.py rollback'
+        """
+        self.setup_stubs()
+        from google.appengine.tools import appcfg
+        appcfg.main(['', 'rollback', '.'])
+
+    def action_prepare(self, options, args):
+        """prepare this project for google appengine.
         """
         from kalapy.conf import settings
         name = settings.PROJECT_NAME
@@ -111,7 +149,10 @@ class GAEProject(ActionCommand):
             print "Creating app.yaml..."
         copy_template('gae_template', os.curdir, context)
 
-    def action_libs(self, options, args):
+        if options.install:
+            self.install_libs(options, args)
+
+    def install_libs(self, options, args):
         """install dependencies in lib dir.
         """
         from werkzeug import import_string
@@ -148,5 +189,4 @@ class GAEProject(ActionCommand):
                 if x in args:
                     os.remove(f)
         os.path.walk("lib", do_clean, ['.pyc', '.pyo', '.so'])
-
